Improve CPU and memory usage of at with negative indexes
Calling `at` with a negative index gets the value from the end of the
iterable.

Previously, this iterated over the whole iterable, put it into an array,
and then used `Array.prototype.at`. This worked, but required storing
the whole iterable in memory.

With this change, we still iterate over the whole iterable. However,
we only need to store a sliding window; if I ask for index -10, I can
discard values "older than" 10.

(In practice, constantly cleaning up this sliding window is slow, so
we only clean it up if its "excess" is too large. I chose a number that
worked well on a simple benchmark on one machine; probably worth
tweaking this number in the future.)